admin: conslidate listing of saml conns and scim dirs into index page
